fix: exception - 예외처리 진행

+++ bookmaru-application/src/main/kotlin/plain/bookmaru/common/command/PageCommand.kt
@@ -4,6 +4,6 @@ data class PageCommand(
     val page: Int,
     val size: Int
 ) {
-    val offset: Int
-        get() = (page * size)
+    val offset: Long
+        get() = (page * size).toLong()
 }
